Refactor test_routine
