+++ tests/test_routine.py
@@ -98,200 +98,6 @@ class TestRoutineMethods(unittest.TestCase):
         self.assertAlmostEqual(self.ss.DCOPF.obj.v, self.ss.DCOPF.obj.v2, places=6)
 
 
-class TestDCOPF(unittest.TestCase):
-    """
-    Test routine `DCOPF`.
-    """
-
-    def setUp(self) -> None:
-        self.ss = ams.load(ams.get_case("5bus/pjm5bus_demo.xlsx"),
-                           setup=True, default_config=True, no_output=True)
-        # decrease load first
-        self.ss.PQ.set(src='p0', attr='v', idx=['PQ_1', 'PQ_2'], value=[0.3, 0.3])
-
-    def test_trip(self):
-        """
-        Test generator trip.
-        """
-        # --- run DCOPF ---
-        self.ss.DCOPF.run(solver='CLARABEL')
-        obj = self.ss.DCOPF.obj.v
-
-        # --- set load ---
-        self.ss.PQ.set(src='p0', attr='v', idx='PQ_1', value=0)
-        self.ss.DCOPF.update()
-
-        self.ss.DCOPF.run(solver='CLARABEL')
-        obj_pqt = self.ss.DCOPF.obj.v
-        self.assertLess(obj_pqt, obj, "Load set does not take effect in DCOPF!")
-
-        # --- trip load ---
-        self.ss.PQ.set(src='u', attr='v', idx='PQ_1', value=0)
-        self.ss.DCOPF.update()
-
-        self.ss.DCOPF.run(solver='CLARABEL')
-        obj_pqt2 = self.ss.DCOPF.obj.v
-        self.assertLess(obj_pqt2, obj_pqt, "Load trip does not take effect in DCOPF!")
-
-        # --- trip generator ---
-        self.ss.StaticGen.set(src='u', attr='v', idx='PV_1', value=0)
-        self.ss.DCOPF.update()
-
-        self.ss.DCOPF.run(solver='CLARABEL')
-        self.assertTrue(self.ss.DCOPF.converged, "DCOPF did not converge under generator trip!")
-        self.assertAlmostEqual(self.ss.DCOPF.get(src='pg', attr='v', idx='PV_1'),
-                               0, places=6,
-                               msg="Generator trip does not take effect in DCOPF!")
-        self.assertAlmostEqual(self.ss.DCOPF.get(src='pg', attr='v', idx='PV_1'),
-                               0, places=6,
-                               msg="Generator trip does not take effect in DCOPF!")
-
-        # --- trip line ---
-        self.ss.Line.set(src='u', attr='v', idx='Line_3', value=0)
-        self.ss.DCOPF.update()
-
-        self.ss.DCOPF.run(solver='CLARABEL')
-        self.assertTrue(self.ss.DCOPF.converged, "DCOPF did not converge under line trip!")
-
-        self.assertAlmostEqual(self.ss.DCOPF.get(src='plf', attr='v', idx='Line_3'),
-                               0, places=6,
-                               msg="Line trip does not take effect in DCOPF!")
-
-
-class TestRTED(unittest.TestCase):
-    """
-    Test methods of `RTED`.
-    """
-
-    def setUp(self) -> None:
-        self.ss = ams.load(ams.get_case("5bus/pjm5bus_demo.xlsx"),
-                           setup=True,
-                           default_config=True,
-                           no_output=True,
-                           )
-        self.ss.RTED.run(solver='CLARABEL')
-
-    def test_dc2ac(self):
-        """
-        Test `RTED.init()` method.
-        """
-        self.ss.RTED.dc2ac()
-        self.assertTrue(self.ss.RTED.converted, "AC conversion failed!")
-        self.assertTrue(self.ss.RTED.exec_time > 0, "Execution time is not greater than 0.")
-
-        stg_idx = self.ss.StaticGen.get_idx()
-        pg_rted = self.ss.RTED.get(src='pg', attr='v', idx=stg_idx)
-        pg_acopf = self.ss.ACOPF.get(src='pg', attr='v', idx=stg_idx)
-        np.testing.assert_almost_equal(pg_rted, pg_acopf, decimal=3)
-
-        bus_idx = self.ss.Bus.get_idx()
-        v_rted = self.ss.RTED.get(src='vBus', attr='v', idx=bus_idx)
-        v_acopf = self.ss.ACOPF.get(src='vBus', attr='v', idx=bus_idx)
-        np.testing.assert_almost_equal(v_rted, v_acopf, decimal=3)
-
-        a_rted = self.ss.RTED.get(src='aBus', attr='v', idx=bus_idx)
-        a_acopf = self.ss.ACOPF.get(src='aBus', attr='v', idx=bus_idx)
-        np.testing.assert_almost_equal(a_rted, a_acopf, decimal=3)
-
-    def test_trip(self):
-        """
-        Test generator trip.
-        """
-        self.ss.RTED.run(solver='CLARABEL')
-        obj = self.ss.RTED.obj.v
-
-        # --- set load ---
-        self.ss.PQ.set(src='p0', attr='v', idx='PQ_1', value=0)
-        self.ss.DCOPF.update()
-
-        self.ss.RTED.run(solver='CLARABEL')
-        obj_pqt = self.ss.RTED.obj.v
-        self.assertLess(obj_pqt, obj, "Load set does not take effect in RTED!")
-
-        # --- trip load ---
-        self.ss.PQ.set(src='u', attr='v', idx='PQ_1', value=0)
-        self.ss.RTED.update()
-
-        self.ss.RTED.run(solver='CLARABEL')
-        obj_pqt2 = self.ss.RTED.obj.v
-        self.assertLess(obj_pqt2, obj_pqt, "Load trip does not take effect in RTED!")
-
-        # --- trip generator ---
-        self.ss.StaticGen.set(src='u', attr='v', idx='PV_1', value=0)
-        self.ss.RTED.update()
-
-        self.ss.RTED.run(solver='CLARABEL')
-        self.assertTrue(self.ss.RTED.converged, "RTED did not converge under generator trip!")
-        self.assertAlmostEqual(self.ss.RTED.get(src='pg', attr='v', idx='PV_1'),
-                               0, places=6,
-                               msg="Generator trip does not take effect in RTED!")
-
-        pg_trip = self.ss.RTED.get(src='pg', attr='v', idx='PV_1')
-        np.testing.assert_almost_equal(pg_trip, 0, decimal=6)
-
-        # --- trip line ---
-        self.ss.Line.set(src='u', attr='v', idx='Line_3', value=0)
-        self.ss.RTED.update()
-
-        self.ss.RTED.run(solver='CLARABEL')
-        self.assertTrue(self.ss.RTED.converged, "RTED did not converge under line trip!")
-        self.assertAlmostEqual(self.ss.RTED.get(src='plf', attr='v', idx='Line_3'),
-                               0, places=6,
-                               msg="Line trip does not take effect in RTED!")
-
-
-class TestED(unittest.TestCase):
-    """
-    Test routine `ED`.
-    """
-
-    def setUp(self) -> None:
-        self.ss = ams.load(ams.get_case("5bus/pjm5bus_demo.xlsx"),
-                           setup=True, default_config=True, no_output=True)
-        # decrease load first
-        self.ss.PQ.set(src='p0', attr='v', idx=['PQ_1', 'PQ_2'], value=[0.3, 0.3])
-
-    def test_trip(self):
-        """
-        Test generator trip.
-        """
-        self.ss.ED.run(solver='CLARABEL')
-        obj = self.ss.ED.obj.v
-
-        # --- trip load ---
-        self.ss.PQ.set(src='p0', attr='v', idx='PQ_1', value=0)
-        self.ss.ED.update()
-
-        self.ss.ED.run(solver='CLARABEL')
-        obj_pqt = self.ss.ED.obj.v
-
-        self.assertLess(obj_pqt, obj, "Load trip does not take effect in ED!")
-
-        # --- trip generator ---
-        self.ss.StaticGen.set(src='u', attr='v', idx='PV_1', value=0)
-        self.ss.ED.update()
-
-        self.ss.ED.run(solver='CLARABEL')
-        self.assertTrue(self.ss.ED.converged, "ED did not converge under generator trip!")
-        obj_gt = self.ss.ED.obj.v
-        self.assertGreater(obj_gt, obj, "Generator trip does not take effect in ED!")
-
-        pg_trip = self.ss.ED.get(src='pg', attr='v', idx='PV_1')
-        np.testing.assert_almost_equal(pg_trip, 0, decimal=6)
-
-        # --- trip line ---
-        self.ss.Line.set(src='u', attr='v', idx='Line_3', value=0)
-        self.ss.ED.update()
-
-        self.ss.ED.run(solver='CLARABEL')
-        self.assertTrue(self.ss.Ed.converged, "ED did not converge under line trip!")
-        obj_lt = self.ss.ED.obj.v
-        self.assertGreater(obj_lt, obj_gt, "Line trip does not take effect in ED!")
-
-        plf_trip = self.ss.ED.get(src='plf', attr='v', idx='Line_3')
-        np.testing.assert_almost_equal(plf_trip, 0, decimal=6)
-
-
 class TestSetOptzValueACOPF(unittest.TestCase):
     """
     Test value settings of `OptzBase` series in `ACOPF`.
